NumberConverter WIP. Added number maps (Ones, Tens, etc.).  Added tests, including ones for future.  Only supports up to 19 and multiples of ten, so far.

+++ @number-to-words-api/Program.cs
@@ -9,7 +9,7 @@ public class Program
 
     app.UseHttpsRedirection();
 
-    app.MapGet("/number-to-words-api/convert-price/{number}", (int number) =>
+    app.MapGet("/number-to-words-api/convert-price/{number}", (double number) =>
     {
       string words = NumberConverter.ConvertPriceToWords(number);
       return new { number, words };
